IoTCloudServices/message_router/code/requirements.txt
requests
paho-mqtt
flask
flask_cors